requirements.txt
# ─── core runtime ───
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20

SQLAlchemy==2.0.38
greenlet==3.1.1

httpx==0.28.1
httpx-sse==0.4.0
anyio==4.8.0
h11==0.14.0

langchain-core==0.3.51
langchain-text-splitters==0.3.8
langchain-openai==0.1.6          # add (pin latest)
langsmith==0.3.8

llama-cloud==0.1.17
llama-cloud-services==0.6.9
llama-index-core==0.12.28
llama-index-readers-file==0.4.7

openai==1.70.0
tiktoken==0.9.0

python-dotenv==1.1.0
aiohttp==3.11.12
aiohappyeyeballs==2.4.6
frozenlist==1.5.0
aiosignal==1.3.2
yarl==1.18.3
multidict==6.1.0
requests==2.32.3
urllib3==2.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
orjson==3.10.15
dirtyjson==1.0.8
lxml==5.3.1
filetype==1.2.0
striprtf==0.0.26

# ─── parsing assets ───
pypdf==5.4.0          # single PDF lib
python-docx==1.1.2
Pillow==10.0.1
fonttools==4.43.0
